backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
gunicorn
django-filter
drf-yasg
django-jazzmin
